+++ src/main/java/org/polyfrost/animatium_legacy/mixin/RenderEntityItemMixin.java
@@ -3,13 +3,20 @@ package org.polyfrost.animatium_legacy.mixin;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.entity.Render;
 import net.minecraft.client.renderer.entity.RenderEntityItem;
+import net.minecraft.client.renderer.entity.RenderItem;
 import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.client.resources.model.IBakedModel;
 import net.minecraft.entity.item.EntityItem;
+import net.minecraft.init.Items;
+import net.minecraft.item.ItemPotion;
+import net.minecraft.item.ItemStack;
 import org.polyfrost.animatium_legacy.Animatium;
 import org.polyfrost.animatium_legacy.config.AnimatiumSettings;
 import org.polyfrost.animatium_legacy.hooks.DroppedItemHook;
+import org.polyfrost.animatium_legacy.init.CustomModelBakery;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -20,25 +27,24 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(RenderEntityItem.class)
 public abstract class RenderEntityItemMixin extends Render<EntityItem> {
-    //    @Shadow @Final private RenderItem itemRenderer;
+    @Shadow
+    @Final
+    private RenderItem itemRenderer;
+
     @Unique
     private boolean animatium$isGui3d;
 
-//    @Unique
-//    private ItemStack animatium$stack = null;
+    @Unique
+    private ItemStack animatium$stack = null;
 
     protected RenderEntityItemMixin(final RenderManager renderManager) {
         super(renderManager);
     }
 
-    @Inject(
-            method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V",
-            at = @At(
-                    value = "HEAD"
-            )
-    )
+    @Inject(method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V", at = @At(value = "HEAD"))
     private void animatium$setHook(final EntityItem entity, final double x, final double y, final double z, final float entityYaw, final float tickDelta, final CallbackInfo ci) {
         DroppedItemHook.isItemDropped = true;
+        this.animatium$stack = entity.getEntityItem();
     }
 
     @Inject(method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V", at = @At(value = "TAIL"))
@@ -48,13 +54,13 @@ public abstract class RenderEntityItemMixin extends Render<EntityItem> {
 
     @ModifyVariable(method = "func_177077_a", at = @At("STORE"), name = "flag")
     private boolean animatium$hookGui3d(final boolean original) {
-        animatium$isGui3d = original;
+        this.animatium$isGui3d = original;
         return original;
     }
 
     @ModifyArg(method = "func_177077_a", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/GlStateManager;rotate(FFFF)V"), index = 0)
     private float animatium$apply2dItem(final float original) {
-        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.itemSprites && !Animatium.isItemPhysics && !animatium$isGui3d) {
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.itemSprites && !Animatium.isItemPhysics && !this.animatium$isGui3d) {
             return 180.0F - renderManager.playerViewY;
         } else {
             return original;
@@ -63,56 +69,30 @@ public abstract class RenderEntityItemMixin extends Render<EntityItem> {
 
     @Inject(method = "func_177077_a", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/GlStateManager;rotate(FFFF)V", shift = At.Shift.AFTER))
     private void animatium$fix2dRotation(final EntityItem itemIn, final double p_177077_2_, final double p_177077_4_, final double p_177077_6_, final float p_177077_8_, final IBakedModel p_177077_9_, final CallbackInfoReturnable<Integer> cir) {
-        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.itemSprites && AnimatiumSettings.rotationFix && !Animatium.isItemPhysics && !animatium$isGui3d) {
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.itemSprites && AnimatiumSettings.rotationFix && !Animatium.isItemPhysics && !this.animatium$isGui3d) {
             GlStateManager.rotate(-renderManager.playerViewX, 1.0F, 0.0F, 0.0F);
         }
     }
 
-//    @ModifyVariable(
-//            method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V",
-//            at = @At(
-//                    value = "STORE"
-//            ),
-//            index = 10
-//    )
-//    private ItemStack animatium$captureStack(ItemStack stack) {
-//        animatium$stack = stack;
-//        return stack;
-//    }
-//
-//    @ModifyArg(
-//            method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V",
-//            at = @At(
-//                    value = "INVOKE",
-//                    target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V"
-//            ),
-//            index = 1
-//    )
-//    private IBakedModel animatium$swapToCustomModel(IBakedModel model) {
-//        if (!OldAnimationsSettings.INSTANCE.enabled || !OldAnimationsSettings.oldPotionsDropped) return model;
-//        if (animatium$stack.getItem() instanceof ItemPotion) {
-//            return CustomModelBakery.BOTTLE_OVERLAY.getBakedModel();
-//        }
-//        return model;
-//    }
-//
-//    @Inject(
-//            method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V",
-//            at = @At(
-//                    value = "INVOKE",
-//                    target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V",
-//                    shift = At.Shift.AFTER
-//            )
-//    )
-//    private void animatium$renderCustomBottle(EntityItem entity, double x, double y, double z, float entityYaw, float tickDelta, CallbackInfo ci) {
-//        if (!OldAnimationsSettings.INSTANCE.enabled || !OldAnimationsSettings.oldPotionsDropped) return;
-//        if (entity.getEntityItem().getItem() instanceof ItemPotion) {
-//            itemRenderer.renderItem(new ItemStack(Items.glass_bottle), animatium$getBottleModel(entity.getEntityItem()));
-//        }
-//    }
-//
-//    @Unique
-//    private IBakedModel animatium$getBottleModel(ItemStack stack) {
-//        return ItemPotion.isSplash(stack.getMetadata()) ? CustomModelBakery.BOTTLE_SPLASH_EMPTY.getBakedModel() : CustomModelBakery.BOTTLE_DRINKABLE_EMPTY.getBakedModel();
-//    }
+    @ModifyArg(method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V"), index = 1)
+    private IBakedModel animatium$swapToCustomModel(final IBakedModel model) {
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.oldPotionsDropped && animatium$stack.getItem() instanceof ItemPotion) {
+            return CustomModelBakery.BOTTLE_OVERLAY.getBakedModel();
+        } else {
+            return model;
+        }
+    }
+
+    @Inject(method = "doRender(Lnet/minecraft/entity/item/EntityItem;DDDFF)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V", shift = At.Shift.AFTER))
+    private void animatium$renderCustomBottle(final EntityItem entity, final double x, final double y, final double z, final float entityYaw, final float tickDelta, final CallbackInfo ci) {
+        final ItemStack stack = entity.getEntityItem();
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.oldPotionsDropped && stack.getItem() instanceof ItemPotion) {
+            IBakedModel potionModel = CustomModelBakery.BOTTLE_DRINKABLE_EMPTY.getBakedModel();
+            if (ItemPotion.isSplash(stack.getMetadata())) {
+                potionModel = CustomModelBakery.BOTTLE_SPLASH_EMPTY.getBakedModel();
+            }
+
+            this.itemRenderer.renderItem(new ItemStack(Items.glass_bottle), potionModel);
+        }
+    }
 }