+++ src/main/java/org/polyfrost/animatium_legacy/mixin/RenderItemMixin.java
@@ -14,6 +14,7 @@ import net.minecraft.client.resources.model.IBakedModel;
 import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.Items;
 import net.minecraft.item.*;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ResourceLocation;
@@ -22,6 +23,7 @@ import org.polyfrost.animatium_legacy.config.AnimatiumSettings;
 import org.polyfrost.animatium_legacy.hooks.DroppedItemHook;
 import org.polyfrost.animatium_legacy.hooks.GlintModelHook;
 import org.polyfrost.animatium_legacy.hooks.TransformTypeHook;
+import org.polyfrost.animatium_legacy.init.CustomModelBakery;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -35,13 +37,18 @@ import java.util.stream.Collectors;
 
 @Mixin(RenderItem.class)
 public abstract class RenderItemMixin {
-    //    @Shadow
-//    public abstract void renderItem(ItemStack stack, IBakedModel model);
+    @Shadow
+    public abstract void renderItem(final ItemStack stack, final IBakedModel model);
+
     @Shadow
     @Final
     private static ResourceLocation RES_ITEM_GLINT;
-    //    @Unique private ItemStack animatium$stackGui = null;
-//    @Unique private ItemStack animatium$stackHeld = null;
+
+    @Unique
+    private ItemStack animatium$stackGui = null;
+
+    @Unique
+    private ItemStack animatium$stackHeld = null;
 
     @Unique
     private ItemStack animatium$stack = null;
@@ -91,8 +98,7 @@ public abstract class RenderItemMixin {
                 ci.cancel();
             }
 
-            if (AnimatiumSettings.enchantmentGlintGui && TransformTypeHook.isRenderingInGUI() && !Animatium.isNEUPresent) {
-//                if (OldAnimationsSettings.oldPotionsGui && animatium$stackGui.getItem() instanceof ItemPotion) { return; }
+            if (AnimatiumSettings.enchantmentGlintGui && TransformTypeHook.isRenderingInGUI() && !Animatium.isNEUPresent && !(AnimatiumSettings.oldPotionsGui && animatium$stackGui.getItem() instanceof ItemPotion)) {
                 ci.cancel();
             }
         }
@@ -188,9 +194,7 @@ public abstract class RenderItemMixin {
 
     @Inject(method = "renderItemIntoGUI", at = @At(value = "TAIL"))
     private void animatium$renderGuiGlint(final ItemStack stack, final int x, final int y, final CallbackInfo ci) {
-        if (AnimatiumSettings.potionGlint && stack.getItem() instanceof ItemPotion) return;
-//        if (OldAnimationsSettings.oldPotionsGui && stack.getItem() instanceof ItemPotion) return;
-        if (AnimatiumSettings.INSTANCE.enabled && !Animatium.isNEUPresent && AnimatiumSettings.enchantmentGlintGui && stack.hasEffect()) {
+        if (AnimatiumSettings.INSTANCE.enabled && !Animatium.isNEUPresent && AnimatiumSettings.enchantmentGlintGui && stack.hasEffect() && !((AnimatiumSettings.potionGlint || AnimatiumSettings.oldPotionsGui) && stack.getItem() instanceof ItemPotion)) {
             GlintModelHook.INSTANCE.renderGlintGui(x, y, RES_ITEM_GLINT);
         }
     }
@@ -228,100 +232,56 @@ public abstract class RenderItemMixin {
         return stack;
     }
 
-//    @ModifyVariable(
-//            method = "renderItemModelTransform",
-//            at = @At(
-//                    value = "HEAD",
-//                    ordinal = 0
-//            ),
-//            index = 1,
-//            argsOnly = true
-//    )
-//    private ItemStack animatium$captureHeldStack(ItemStack stack) {
-//        animatium$stackHeld = stack;
-//        return stack;
-//    }
-//
-//    @ModifyArg(
-//            method = "renderItemModelTransform",
-//            at = @At(
-//                    value = "INVOKE",
-//                    target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V"
-//            ),
-//            index = 1
-//    )
-//    private IBakedModel animatium$swapToCustomModel(IBakedModel model) {
-//        if (!OldAnimationsSettings.INSTANCE.enabled || !OldAnimationsSettings.oldPotions) return model;
-//        if (animatium$stackHeld.getItem() instanceof ItemPotion) {
-//            return CustomModelBakery.BOTTLE_OVERLAY.getBakedModel();
-//        }
-//        return model;
-//    }
-//
-//    @Inject(
-//            method = "renderItemModelTransform",
-//            at = @At(
-//                    value = "INVOKE",
-//                    target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V",
-//                    shift = At.Shift.AFTER
-//            )
-//    )
-//    private void animatium$renderCustomBottle(ItemStack stack, IBakedModel model, ItemCameraTransforms.TransformType cameraTransformType, CallbackInfo ci) {
-//        if (!OldAnimationsSettings.INSTANCE.enabled || !OldAnimationsSettings.oldPotions) return;
-//
-//        if (stack.getItem() instanceof ItemPotion) {
-//            renderItem(new ItemStack(Items.glass_bottle), animatium$getBottleModel(stack));
-//        }
-//    }
-//
-//    @ModifyVariable(
-//            method = "renderItemIntoGUI",
-//            at = @At(
-//                    value = "HEAD",
-//                    ordinal = 0
-//            ),
-//            index = 1,
-//            argsOnly = true
-//    )
-//    private ItemStack animatium$captureGuiStack(ItemStack stack) {
-//        animatium$stackGui = stack;
-//        return stack;
-//    }
-//
-//    @ModifyArg(
-//            method = "renderItemIntoGUI",
-//            at = @At(
-//                    value = "INVOKE",
-//                    target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V"
-//            ),
-//            index = 1
-//    )
-//    private IBakedModel animatium$swapToCustomModel2(IBakedModel model) {
-//        if (!OldAnimationsSettings.INSTANCE.enabled || !OldAnimationsSettings.oldPotionsGui) return model;
-//        if (animatium$stackGui.getItem() instanceof ItemPotion) {
-//            return CustomModelBakery.BOTTLE_OVERLAY.getBakedModel();
-//        }
-//        return model;
-//    }
-//
-//    @Inject(
-//            method = "renderItemIntoGUI",
-//            at = @At(
-//                    value = "INVOKE",
-//                    target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V",
-//                    shift = At.Shift.AFTER
-//            )
-//    )
-//    private void animatium$renderCustomBottle2(ItemStack stack, int x, int y, CallbackInfo ci) {
-//        if (!OldAnimationsSettings.INSTANCE.enabled || !OldAnimationsSettings.oldPotionsGui) return;
-//
-//        if (stack.getItem() instanceof ItemPotion) {
-//            renderItem(new ItemStack(Items.glass_bottle), animatium$getBottleModel(stack));
-//        }
-//    }
-//
-//    @Unique
-//    private IBakedModel animatium$getBottleModel(ItemStack stack) {
-//        return ItemPotion.isSplash(stack.getMetadata()) ? CustomModelBakery.BOTTLE_SPLASH_EMPTY.getBakedModel() : CustomModelBakery.BOTTLE_DRINKABLE_EMPTY.getBakedModel();
-//    }
+    @ModifyVariable(method = "renderItemModelTransform", at = @At(value = "HEAD", ordinal = 0), index = 1, argsOnly = true)
+    private ItemStack animatium$captureHeldStack(final ItemStack stack) {
+        this.animatium$stackHeld = stack;
+        return stack;
+    }
+
+    @ModifyArg(method = "renderItemModelTransform", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V"), index = 1)
+    private IBakedModel animatium$swapToCustomModel(final IBakedModel model) {
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.oldPotions && this.animatium$stackHeld.getItem() instanceof ItemPotion) {
+            return CustomModelBakery.BOTTLE_OVERLAY.getBakedModel();
+        } else {
+            return model;
+        }
+    }
+
+    @Inject(method = "renderItemModelTransform", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V", shift = At.Shift.AFTER))
+    private void animatium$renderCustomBottle(final ItemStack stack, final IBakedModel model, final ItemCameraTransforms.TransformType cameraTransformType, final CallbackInfo ci) {
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.oldPotions && stack.getItem() instanceof ItemPotion) {
+            this.renderItem(new ItemStack(Items.glass_bottle), animatium$getBottleModel(stack));
+        }
+    }
+
+    @ModifyVariable(method = "renderItemIntoGUI", at = @At(value = "HEAD", ordinal = 0), index = 1, argsOnly = true)
+    private ItemStack animatium$captureGuiStack(final ItemStack stack) {
+        this.animatium$stackGui = stack;
+        return stack;
+    }
+
+    @ModifyArg(method = "renderItemIntoGUI", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V"), index = 1)
+    private IBakedModel animatium$swapToCustomModel2(IBakedModel model) {
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.oldPotionsGui && this.animatium$stackGui.getItem() instanceof ItemPotion) {
+            return CustomModelBakery.BOTTLE_OVERLAY.getBakedModel();
+        } else {
+            return model;
+        }
+    }
+
+    @Inject(method = "renderItemIntoGUI", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/RenderItem;renderItem(Lnet/minecraft/item/ItemStack;Lnet/minecraft/client/resources/model/IBakedModel;)V", shift = At.Shift.AFTER))
+    private void animatium$renderCustomBottleGui(ItemStack stack, int x, int y, CallbackInfo ci) {
+        if (AnimatiumSettings.INSTANCE.enabled && AnimatiumSettings.oldPotionsGui && stack.getItem() instanceof ItemPotion) {
+            this.renderItem(new ItemStack(Items.glass_bottle), animatium$getBottleModel(stack));
+        }
+    }
+
+    @Unique
+    private IBakedModel animatium$getBottleModel(final ItemStack stack) {
+        if (ItemPotion.isSplash(stack.getMetadata())) {
+            return CustomModelBakery.BOTTLE_SPLASH_EMPTY.getBakedModel();
+        } else {
+            return CustomModelBakery.BOTTLE_DRINKABLE_EMPTY.getBakedModel();
+        }
+    }
 }